P0 verification + post-hardware fixes
Hardware verification passed on M5StickS3 (/dev/cu.usbmodem21101):
  - LCD shows "Pocket Oracle"
  - USB-CDC console responds at 115200 baud, "pocket> " prompt
  - help / version commands work

Fixes discovered during bring-up:
  - Vendor in M5Unified 0.2.14 + M5GFX 0.2.20 as git submodules
    (not on the ESP Component Registry). Pinned to same tags as
    atoms3r-home-assistant for proven IDF 5.4.2 compatibility.
  - Disable WiFi RX buffer overrides (defaults collide with BA window).
  - Disable CONFIG_PM_ENABLE / tickless idle for P0..P7.
    Auto light-sleep tears down USB-Serial-JTAG enumeration; will
    re-enable in P8 with PM locks around the active console + UI path.
  - GCC 14 warning fixes:
    * project-wide -Wno-missing-field-initializers (M5Unified/M5GFX style)
    * per-target -Wno-error=maybe-uninitialized on vendor components
    * static_cast<int>() on M5GFX width()/height() in ESP_LOGI args
  - Drop m5stack/M5Unified from main/idf_component.yml; require M5Unified
    component name directly in main/CMakeLists.txt.

Correct KEY1 / KEY2 mapping in docs:
  - KEY1 (G11) = side small button, PMIC power/Boot — NOT a free user key.
    Single press = power on/wake, double press = power off, hold = bootloader.
  - KEY2 (G12) = front big button = the primary user-facing button.
  Updated README.md, PRD.md, CLAUDE.md, docs/HARDWARE_M5STICKS3.md.

+++ firmware/main/main.cc
@@ -41,7 +41,8 @@ extern "C" void app_main(void)
                           M5.Display.height() / 2);
 
     ESP_LOGI(TAG, "Pocket Oracle booted — display %dx%d",
-             M5.Display.width(), M5.Display.height());
+             static_cast<int>(M5.Display.width()),
+             static_cast<int>(M5.Display.height()));
 
     // USB-CDC service console (B2: enabled by default during development).
     pocket::console_start();
